make posts + static controllers prettier

+++ app/controllers/static_controller.rb
@@ -17,7 +17,7 @@ class StaticController < ApplicationController
   end
 
   def logout
-    session.delete(:user_id)
+    session.delete(:user_id) if session[:user_id]
     redirect_to root_path
   end
